data/other/requirements.txt
absl-py==1.2.0
addict==2.4.0
basicsr==1.3.3.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.2
charset-normalizer==2.0.6
cloudpickle==2.0.0
cmake==3.18.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
future==0.18.2
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
h5py==3.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.3
intel-openmp==2019.0
joblib==1.0.1
kiwisolver==1.1.0
lmdb==1.3.0
Markdown==3.3.7
matplotlib==3.0.3
mkl==2019.0
mkl-include==2019.0
networkx==2.4
numpy==1.18.5
oauthlib==3.2.1
opencv-python==4.6.0.66
Pillow==7.2.0
progressbar2==3.53.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.15.0
scikit-learn==0.24.2
scipy==1.4.1
six==1.15.0
tb-nightly==2.11.0a20220816
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorlayer==2.2.3
threadpoolctl==2.2.0
tifffile==2020.9.3
timm==0.6.7
torch==1.10.1
torchinfo==1.5.4
torchsummary==1.5.1
torchvision==0.2.1
tqdm==4.19.9
typing==3.7.4.3
typing_extensions==4.1.1
urllib3==1.26.7
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.14.1
yapf==0.32.0
zipp==3.6.0
